# docs/swagger.yaml
definitions:
  memorize_internal_service_authentication.LoginUserRequest:
    properties:
      password:
        maxLength: 70
        type: string
      username:
        maxLength: 100
        type: string
    required:
    - password
    - username
    type: object
  memorize_internal_service_authentication.RegisterUserRequest:
    properties:
      email:
        maxLength: 100
        type: string
      password:
        maxLength: 70
        type: string
    required:
    - email
    - password
    type: object
  memorize_internal_service_authentication_user.CreateUserRequest:
    properties:
      email:
        maxLength: 100
        type: string
      password:
        maxLength: 70
        type: string
      username:
        maxLength: 100
        type: string
    required:
    - email
    - password
    - username
    type: object
  memorize_internal_service_authentication_user.UpdateUserRequest:
    properties:
      email:
        maxLength: 100
        type: string
      password:
        maxLength: 70
        type: string
      username:
        maxLength: 100
        type: string
    required:
    - email
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /admin/users:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      responses: {}
      tags:
      - Users
    post:
      parameters:
      - description: entity to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/memorize_internal_service_authentication_user.CreateUserRequest'
      responses: {}
      tags:
      - Users
  /admin/users/{id}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      tags:
      - Users
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      tags:
      - Users
    put:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: entity to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/memorize_internal_service_authentication_user.UpdateUserRequest'
      responses: {}
      tags:
      - Users
  /auth/login:
    post:
      parameters:
      - description: login
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/memorize_internal_service_authentication.LoginUserRequest'
      responses: {}
      tags:
      - Authentication
  /auth/register:
    post:
      parameters:
      - description: register
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/memorize_internal_service_authentication.RegisterUserRequest'
      responses: {}
      tags:
      - Authentication
swagger: "2.0"